+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 =========
 # All notable changes to this project will be documented in this file.
 
+## 2.0.0 - 2020-05-10
+### Added
+* Scalar types validator
+
+### Updated
+* Class validator allows only one class or interfce name as type to checks insted of many items
+* Changelog file
+* Readme file
+
 ## 1.0.0 - 2020-05-07
 ### Added
 * This changelog file